simulator/requirements.txt
kafka-python==2.2.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
